멤버목록 이미지 컴포넌트 분리
